+++ CHANGELOG.md
@@ -2,4 +2,5 @@ Please clear changelog after each release.
 Put the changelog BELOW the dashes. ANYTHING ABOVE IS IGNORED.
 -----------------
 - Updated `InterpolatedGravityFunction` to now properly supply a bottom and top gravity value.
+- Multiple `GravityBelt`s can now apply at the same position.
 - Fixed a crash upon booting servers.